backend/requirements.txt
# 基础依赖
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.8

# 数据库相关
SQLAlchemy==2.0.40
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0

# 安全认证
PyJWT==2.10.1
python-dotenv==1.1.0

# 数据验证
pydantic[email]==2.11.3

# 生产环境
gunicorn==23.0.0

# 开发环境
flask-debugtoolbar==0.16.0
python-dotenv==1.1.0